fix: user dto naming and divided into files
