+++ server/app.controller.ts
@@ -8,13 +8,14 @@ import { LocalAuthGuard } from "./routes/auth/guard/local-auth.guard";
 import { RolesGuard } from "./routes/auth/guard/roles.guard";
 import { Roles } from "./routes/auth/roles/roles.decorator";
 import { Role } from "./routes/auth/roles/roles.enum";
-import { CreateUserDto, FindUserDto } from "./routes/users/dto/users.dto";
+import { CreateUserDto } from "./routes/users/dto/create-user.dto";
+import { FindUserResponseDto } from "routes/users/dto/find-user.dto";
 
 @Controller()
 export class AppController {
   constructor(private readonly appService: AppService, private readonly authService: AuthService) {}
 
-  @ApiBody({ type: FindUserDto })
+  @ApiBody({ type: FindUserResponseDto })
   @UseGuards(LocalAuthGuard)
   @Post("login")
   login(@Req() req: Request, @Res({ passthrough: true }) res: Response): void {
