+++ server/routes/users/dto/update-user.dto.ts
@@ -1,0 +1,16 @@
+import { ApiProperty } from "@nestjs/swagger";
+import { UpdateUserDto as UpdateUserDtoType } from "@/dto";
+
+export class UpdateUserDto implements UpdateUserDtoType {
+  @ApiProperty()
+  email?: string;
+
+  @ApiProperty()
+  role?: "user" | "admin";
+
+  @ApiProperty()
+  isActive?: boolean;
+
+  @ApiProperty()
+  profileImgUrl?: string;
+}
